LLM_bot/requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiogram_tests @ git+https://github.com/Like6po/aiogram_tests.git@3574eb83b3fa3c98496a5141b9fa491054c09b04
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
asyncio==3.4.3
attrs==23.2.0
boto3==1.34.14
botocore==1.34.14
certifi==2023.11.17
charset-normalizer==3.3.2
coverage==7.4.1
diskcache==5.6.3
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
huggingface-hub==0.20.2
idna==3.6
iniconfig==2.0.0
install==1.3.5
Jinja2==3.1.2
jmespath==1.0.1
llama_cpp_python==0.2.27
logging==0.4.9.6
magic-filter==1.0.12
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pluggy==1.4.0
pydantic==2.5.3
pydantic_core==2.14.6
pytest==8.0.0
pytest-asyncio==0.23.5
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
s3transfer==0.10.0
safetensors==0.4.1
six==1.16.0
sympy==1.12
tokenizers==0.15.0
torch==2.1.2
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.9.0
urllib3==2.0.7
yarl==1.9.4
